Refactoring code for consistency
