+++ src/components/App/index.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import Main from '../Main';
 
-function App() {
+const App: React.FC = () => {
     return (
         <>
             <Main />
         </>
     );
-}
+};
 
 export default App;
