+++ src/components/InfiniteScroll/index.tsx
@@ -1,0 +1,3 @@
+import InfiniteScroll from './InfiniteScroll';
+
+export default InfiniteScroll;
